buildsys: workaround macOS craziness
After editing the nq source code and running `make`, on macOS 10.15 or
later the `bin/GAPARCH/nq` binary would crash upon launch, with just the
message `Killed: 9`.

Apparently this is caused by macOS’s linkage of codesign hashes to inode
numbers; once an executable has been loaded, the inode<->codesign hash cache
gets filled, and it cannot be emptied until the inode is deleted or the system
is restarted. This means that if you have a build script that overwrites a
binary (instead of deleting and then recreating the file) any operation to
execute it will fail like this.